Added the calc_log_parabolic_spectrum_max_freq to a new analysis module
