+++ docs/examples.rst
@@ -148,7 +148,7 @@ You can use the pulsar's fit to calculate the peak frequency like so:
 
     from pulsar_spectra.spectral_fit import find_best_spectral_fit
     from pulsar_spectra.catalogue import collect_catalogue_fluxes
-    from pulsar_spectra.models import calc_log_parabolic_spectrum_max_freq
+    from pulsar_spectra.analysis import calc_log_parabolic_spectrum_max_freq
 
     cat_dict = collect_catalogue_fluxes()
     pulsar = 'J1136+1551'
@@ -178,7 +178,7 @@ Estimate emission height from a high-frequency cut-off power-law fit
 
 As demonstrated in Jankowski et al. (2018), we can use the high-frequency cut-off power-law model
 from Kontorovich & Flanchick (2013) to estimate the location of the centre of the magnetic polar cap,
-assuming a canonical neutron star (radius of 12+/-2 km; Steiner et al., 2018) and a dipole magnetic field. 
+assuming a canonical neutron star (radius of 12+/-2 km; Steiner et al., 2018) and a dipole magnetic field.
 To perform this calculation, use the in-built function as follows:
 
 .. script location: example_scripts/estimate_emission_height.py
@@ -194,7 +194,7 @@ To perform this calculation, use the in-built function as follows:
     model_name, m, _, _, _ = find_best_spectral_fit(pulsar, freqs, fluxs, flux_errs, refs)
     if model_name == "high_frequency_cut_off_power_law":
         B_pc, u_B_pc, B_surf, B_lc, r_lc, z_e, u_z_e, z_percent, u_z_percent = calc_high_frequency_cutoff_emission_height(
-            pulsar, 
+            pulsar,
             m.values[0],
             m.errors[0],
         )
